[FE] Refactor : image API요청, Editor 파일 분리
 image API요청하는 부분과 Editor가 정의되어져 있는 파일을 분리했습니다.
